Module scrypto_statictypes::prelude

Convenience re-export of common members

Like the standard library’s prelude, this module simplifies importing of common items. Unlike the standard prelude, the contents of this module must be imported manually:

use scrypto_statictypes::prelude::*;

Re-exports

pub use crate::declare_resource;

pub use crate::XRD;

Macros

declare_resource

Structs

BucketOf

ProofOf

A custom ProofOf using RefCell so we can implement Drop

ResourceAddress

Represents a resource address.

ResourceOf

VaultOf

Traits

Deposit

DepositOf

DepositOfExplicit

Explicitly requires deposit_of:: syntax instead of of automatically allowing any BucketOf<_> parameter

StaticResource

StaticResourceDecl

UncheckedInto

UncheckedIntoProofOf

Unwrap

WithInner

Withdraw

WithdrawOf

WithdrawOfExplicit

Explicitly requires withdraw_of:: syntax instead of of automatically allowing any ResourceOf<_> parameter