Struct scrypto_statictypes::prelude::VaultOf

pub struct VaultOf<RES> { /* fields omitted */ }

VaultOf

Implementations

impl<RES: Resource> VaultOf<RES>

pub fn new<A: Into<ResourceDef>>(resource_def: A) -> Self

impl<RES> VaultOf<RES>

pub fn with_bucket(bucketof: BucketOf<RES>) -> VaultOf<RES>

pub fn put(&self, other: BucketOf<RES>)

Puts a typed bucket of resources into this vault.

pub fn take<A: Into<Decimal>>(&self, amount: A) -> BucketOf<RES>

Takes some amount of resources out of this vault, with typed result.

pub fn take_all(&self) -> BucketOf<RES>

Takes all resourced stored in this vault, with typed result.